Renamed components: pr comment fixes
