+++ client/src/Components/Navigation/Navigation.js
@@ -1,84 +1,127 @@
-import React from 'react';
-import { Switch, Redirect } from 'react-router-dom';
+import React, { useState, useEffect, Fragment } from 'react';
+import { useHistory } from 'react-router-dom';
+import Slide from '@material-ui/core/Slide';
+import Toolbar from '@material-ui/core/Toolbar';
+import SwipeableDrawer from '@material-ui/core/SwipeableDrawer';
+import useScrollTrigger from '@material-ui/core/useScrollTrigger';
 
-import { isLoggedIn } from '../../logic/auth';
+import { logOut } from '../../logic/auth';
+import { getUser, getNotifications } from '../../logic/api';
+import { classifyNotifications } from '../../logic/utils';
 
-import {
-  PublicRoute,
-  PrivateRoute,
-  AdminRoute,
-  AnyRoute,
-} from '../Routes/Routes';
+import { useUser, useSetUser } from '../UserProvider/UserProvider';
+import { useSetNotifications } from '../NotificationsProvider/NotificationsProvider';
 
-import Welcome from '../Welcome/Welcome';
-import LogInForm from '../LogInForm/LogInForm';
-import SignUpForm from '../SignUpForm/SignUpForm';
+import i18n from '../../i18n';
+import ErrorSnackbar from '../ErrorSnackbar/ErrorSnackbar';
+import NavigationBar from '../NavigationBar/NavigationBar';
+import NavigationDrawer from '../NavigationDrawer/NavigationDrawer';
 
-import Home from '../Home/Home';
-import Events from '../Events/Events';
-import UserEvents from '../UserEvents/UserEvents';
-import EventSteps from '../EventSteps/EventSteps';
-import Event from '../Event/Event';
-import Users from '../Users/Users';
-import User from '../User/User';
-import UserTabs from '../UserTabs/UserTabs';
-import Interests from '../Interests/Interests';
-import Chats from '../Chats/Chats';
-import Chat from '../Chat/Chat';
-import Notifications from '../Notifications/Notifications';
-import Email from '../Email/Email';
+import Style from '../Style/Style';
 
-import AdminDashboard from '../AdminDashboard/AdminDashboard';
+const HideOnScroll = (props) => {
+  const { children } = props;
+  const trigger = useScrollTrigger();
 
-import About from '../About/About';
-
-const Navigation = () => {
   return (
-    <Switch>
-      <PublicRoute exact path="/welcome" component={Welcome} />
-      <PublicRoute exact path="/log-in" component={LogInForm} />
-      <PublicRoute exact path="/sign-up" component={SignUpForm} />
+    <Slide appear={false} direction="down" in={!trigger}>
+      {children}
+    </Slide>
+  );
+};
 
-      <PrivateRoute exact path="/" component={Home} />
-      <PrivateRoute exact path="/events" component={Events} />
-      <PrivateRoute exact path="/events/mine" component={UserEvents} />
-      <PrivateRoute exact path="/events/new" component={EventSteps} />
-      <PrivateRoute
-        exact
-        path={['/events/:eventId', '/events/:eventId/comments']}
-        component={Event}
-      />
-      <PrivateRoute exact path="/events/:eventId/edit" component={EventSteps} />
-      <PrivateRoute exact path="/users" component={Users} />
-      <PrivateRoute
-        exact
-        path={['/users/:userId', '/users/:userId/friendships']}
-        component={User}
-      />
-      <PrivateRoute
-        exact
-        path={['/users/:userId/edit', '/users/:userId/edit/interests']}
-        component={UserTabs}
-      />
-      <PrivateRoute exact path="/interests" component={Interests} />
-      <PrivateRoute exact path="/chats" component={Chats} />
-      <PrivateRoute exact path="/chats/:friendshipId" component={Chat} />
-      <PrivateRoute exact path="/notifications" component={Notifications} />
-      <PrivateRoute
-        exact
-        path={[
-          '/users/:userId/confirm-email',
-          '/users/:userId/send-email-confirmation-email',
-        ]}
-        component={Email}
-      />
+const Navigation = (props) => {
+  const { container } = props;
 
-      <AdminRoute exact path="/admin" component={AdminDashboard} />
+  const style = Style();
+  const user = useUser();
+  const setUser = useSetUser();
+  const history = useHistory();
+  const setNotifications = useSetNotifications();
 
-      <AnyRoute exact path="/about" component={About} />
+  const [error, setError] = useState(null);
+  const [loading, setLoading] = useState(true);
+  const [drawerOpen, setDrawerOpen] = useState(false);
+  const [logUserOut, setLogUserOut] = useState(false);
 
-      <Redirect to={isLoggedIn() ? '/' : '/welcome'} />
-    </Switch>
+  useEffect(() => {
+    setLoading(true);
+    setError(null);
+    getUser()
+      .then((data) => {
+        if (data.error || data.errors) {
+          setLogUserOut(true);
+        } else {
+          setUser(data.user);
+          setLoading(false);
+        }
+      })
+      .catch((error) => {
+        setError(error);
+        setLoading(false);
+      });
+  }, [setUser]);
+
+  useEffect(() => {
+    if (user) {
+      setLoading(true);
+      setError(null);
+
+      if (user.preferredLocale !== i18n.language) {
+        i18n.changeLanguage(user.preferredLocale);
+      }
+
+      getNotifications()
+        .then((data) =>
+          setNotifications(classifyNotifications(data.notifications))
+        )
+        .catch((error) => setError(error))
+        .finally(() => setLoading(false));
+    }
+  }, [user, setNotifications]);
+
+  useEffect(() => {
+    if (logUserOut) {
+      setUser(null);
+      logOut();
+      history.push('/welcome');
+    }
+  }, [logUserOut, setUser, history]);
+
+  const handleDrawerToggle = () => setDrawerOpen(!drawerOpen);
+
+  const drawer = user && (
+    <NavigationDrawer
+      clickHandler={handleDrawerToggle}
+      logOutHandler={() => setLogUserOut(true)}
+    />
+  );
+
+  return (
+    <Fragment>
+      <HideOnScroll {...props}>
+        <NavigationBar toggleDrawer={handleDrawerToggle} />
+      </HideOnScroll>
+      <Toolbar />
+      {!loading && user && (
+        <SwipeableDrawer
+          anchor="left"
+          container={container}
+          open={drawerOpen}
+          onOpen={handleDrawerToggle}
+          onClose={handleDrawerToggle}
+          classes={{
+            paper: style.drawerPaper,
+          }}
+          ModalProps={{
+            keepMounted: true,
+          }}
+        >
+          {drawer}
+        </SwipeableDrawer>
+      )}
+      {error && <ErrorSnackbar error={error} />}
+    </Fragment>
   );
 };
